+++ src/main/kotlin/vn/com/ntq/nxdev/actions/PromptAction.kt
@@ -6,10 +6,16 @@ import com.google.gson.Gson
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.ui.Messages
 import com.intellij.openapi.wm.ToolWindowManager
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+import okhttp3.*
 import org.markdown4j.Markdown4jProcessor
 import vn.com.ntq.nxdev.settings.MyPluginSettings
+import java.io.IOException
 import java.net.HttpURLConnection
 import java.net.URL
 import javax.swing.SwingUtilities
@@ -18,9 +24,11 @@ open class PromptAction : AnAction() {
 
     val processor = Markdown4jProcessor()
     override fun actionPerformed(e: AnActionEvent) {
-        val question = getQuestion(e)
-        if (question.isNotEmpty()) {
-            showResponse(e, question)
+        GlobalScope.launch(Dispatchers.IO) {
+            val question = getQuestion(e)
+            if (question.isNotEmpty()) {
+                showResponse(e, question)
+            }
         }
     }
      protected open fun getQuestion(e: AnActionEvent): String {
@@ -28,7 +36,10 @@ open class PromptAction : AnAction() {
         val editor = e.getData(PlatformDataKeys.EDITOR)
         if (project != null && editor != null) {
             val selectionModel = editor.selectionModel
-            val selectedText = selectionModel.selectedText
+            var selectedText = "";
+            ApplicationManager.getApplication().runReadAction {
+                selectedText = selectionModel.selectedText.toString()
+            }
             if (!selectedText.isNullOrEmpty()) {
                 return getPrefix() + selectedText
             } else {
@@ -41,19 +52,36 @@ open class PromptAction : AnAction() {
     protected open fun showResponse(e: AnActionEvent, question: String) {
         val project = e.project
         if (project != null) {
-            // Call to external API
-            val response = sendRequest(question)
-            SwingUtilities.invokeLater {
-                val toolWindowManager = ToolWindowManager.getInstance(project)
-                val toolWindow = toolWindowManager.getToolWindow("NxDev") // replace with your tool window id
-                val contentManager = toolWindow?.contentManager
-                val content = contentManager?.getContent(0)?.component
-                if (content is NxDevWindowFactory.NxDevWindows) {
+            val toolWindowManager = ToolWindowManager.getInstance(project)
+            val toolWindow = toolWindowManager.getToolWindow("NxDev") // replace with your tool window id
+            val contentManager = toolWindow?.contentManager
+            val content = contentManager?.getContent(0)?.component
+            if (content is NxDevWindowFactory.NxDevWindows) {
+                SwingUtilities.invokeLater {
                     toolWindow.show();
                     content.requestField.text = getPrefix()
-                    val markdownContent = processor.process(response.choices[0].message.content)
-                    content.responseArea.text = markdownContent
+                    content.responseArea.text = " "
                 }
+            }
+            // Call to external API
+            val response = sendEventStreamRequest(question)
+            response?.body()?.source()?.let {source ->
+                while (true) {
+                    val event = source.readUtf8Line() ?: break
+                    if (event.isBlank())
+                        continue;
+                    val eventJSON = event.removePrefix("data:")
+                    val data = Gson().fromJson(eventJSON, ChatCompletion::class.java);
+
+                    SwingUtilities.invokeLater {
+                        if (content is NxDevWindowFactory.NxDevWindows) {
+                            val sd = content.responseArea.styledDocument;
+                            sd.insertString(sd.length, data?.choices?.getOrNull(0)?.delta?.content?:"", null)
+                        }
+                    }
+
+                }
+
             }
         }
     }
@@ -61,40 +89,45 @@ open class PromptAction : AnAction() {
     protected open fun getPrefix(): String{
         return "";
     }
-    private fun sendRequest(message: String): ChatCompletion {
+
+
+    private fun sendEventStreamRequest(message: String): Response? {
         val url = URL("https://genapi.ntq.ai/v1/chat/completions")
-        val conn = url.openConnection() as HttpURLConnection
-        conn.requestMethod = "POST"
-        conn.setRequestProperty("Content-Type", "application/json")
-        var authorizationKey = "Bearer " + MyPluginSettings.getInstance().apiKey;
-        conn.setRequestProperty("Authorization", authorizationKey)
-        conn.doOutput = true
+        val requestBody = RequestBody.create(
+            MediaType.parse("application/json"), Gson().toJson(
+                PromptAction.JsonRequest(
+                    model = "ntq-coder",
+                    messages = listOf(PromptAction.Message("user", message)),
+                    max_tokens = 4096
+                )
+            ))
 
-        val jsonRequest = JsonRequest(
-                model = "ntq-coder",
-                messages = listOf(Message("user", message)),
-                max_tokens = 4096
-        )
+        val request = Request.Builder()
+            .url(url)
+            .post(requestBody)
+            .addHeader("Content-Type", "application/json")
+            .addHeader("Authorization", "Bearer ${MyPluginSettings.getInstance().apiKey}")
+            .build()
 
-        val requestBody = Gson().toJson(jsonRequest)
-        conn.outputStream.write(requestBody.toByteArray())
+        return OkHttpClient.Builder().connectionSpecs(createConnectionSpecs()).build().newCall(request).execute()
+    }
 
-        val responseCode = conn.responseCode
-        val response = if (responseCode == HttpURLConnection.HTTP_OK) {
-            conn.inputStream.reader().readText()
-        } else {
-            "Error: $responseCode"
-        }
-
-        conn.disconnect()
-        return Gson().fromJson(response, ChatCompletion::class.java)
+    private fun createConnectionSpecs(): List<ConnectionSpec>? {
+        val spec = ConnectionSpec.Builder(ConnectionSpec.COMPATIBLE_TLS)
+            .tlsVersions(TlsVersion.TLS_1_2)
+            .cipherSuites(
+                CipherSuite.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
+            )
+            .build()
+        return listOf<ConnectionSpec>(spec)
     }
 
     data class JsonRequest(
             val model: String,
             val messages: List<Message>,
             val max_tokens: Int,
-            val platform: String = "Intellij"
+            val platform: String = "Intellij",
+            val stream: Boolean = true
     )
 
     data class Message(