+++ src/main/kotlin/vn/com/ntq/nxdev/toolWindow/ChatCompletion.kt
@@ -1,27 +1,24 @@
 package vn.com.ntq.nxdev.toolWindow
 
+import com.google.gson.annotations.JsonAdapter
+
+data class Data(
+        val data: ChatCompletion
+)
 data class ChatCompletion(
         val id: String,
         val objectName: String,
         val created: Int,
         val model: String,
-        val choices: List<Choice>,
-        val usage: Usage
+        val choices: List<Choice>
 )
-
 data class Choice(
         val index: Int,
-        val message: Message,
+        val delta: Delta,
         val finishReason: String?
 )
 
-data class Message(
+data class Delta(
         val role: String,
         val content: String
-)
-
-data class Usage(
-        val promptTokens: Int,
-        val completionTokens: Int,
-        val totalTokens: Int
 )