+++ src/main/kotlin/vn/com/ntq/nxdev/toolWindow/NxDevWindowFactory.kt
@@ -1,25 +1,27 @@
 package vn.com.ntq.nxdev.toolWindow
 
-import vn.com.ntq.nxdev.actions.PromptAction
+import com.google.gson.Gson
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory
-import vn.com.ntq.nxdev.services.MyProjectService
-import com.google.gson.Gson
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
+import okhttp3.*
 import org.markdown4j.Markdown4jProcessor
+import vn.com.ntq.nxdev.actions.PromptAction
+import vn.com.ntq.nxdev.services.MyProjectService
 import vn.com.ntq.nxdev.settings.MyPluginSettings
 import java.awt.BorderLayout
 import java.awt.event.ActionEvent
 import java.awt.event.ActionListener
-import java.net.HttpURLConnection
+import java.io.IOException
 import java.net.URL
 import javax.swing.*
+import javax.swing.text.StyledDocument
 
 
 class NxDevWindowFactory : ToolWindowFactory {
@@ -40,7 +42,7 @@ class NxDevWindowFactory : ToolWindowFactory {
 
         private val service = toolWindow.project.service<MyProjectService>()
         val requestField = JTextField("What do you want to ask?", 50)
-//        val responseArea = JTextArea(5, 50)
+//        val responseArea = JTextArea(5, 50)```
         val responseArea = JTextPane()
         val processor = Markdown4jProcessor()
 
@@ -60,42 +62,84 @@ class NxDevWindowFactory : ToolWindowFactory {
             override fun actionPerformed(e: ActionEvent) {
                 val request = requestField.text
                 GlobalScope.launch(Dispatchers.IO) {
-                    val response = sendRequest(request)
-                    SwingUtilities.invokeLater {
-                        val content = processor.process(response.choices[0].message.content)
-                        responseArea.text =  content
+//                    val response = sendEventStreamRequest(request)
+                    val url = URL("https://genapi.ntq.ai/v1/chat/completions")
+                    val requestBody = RequestBody.create(
+                        MediaType.parse("application/json"), Gson().toJson(
+                            PromptAction.JsonRequest(
+                                model = "ntq-coder",
+                                messages = listOf(PromptAction.Message("user", request)),
+                                max_tokens = 4096
+                            )
+                        ))
+
+                    val httpRequest = Request.Builder()
+                        .url(url)
+                        .post(requestBody)
+                        .addHeader("Content-Type", "application/json")
+                        .addHeader("Authorization", "Bearer ${MyPluginSettings.getInstance().apiKey}")
+                        .build()
+
+                    val response = OkHttpClient.Builder().connectionSpecs(createConnectionSpecs()).build().newCall(httpRequest).execute()
+                    responseArea.text = " ";
+                    val sd = responseArea.styledDocument;
+
+                    response.body()?.source()?.let {source ->
+                        while (true) {
+                            val event = source.readUtf8Line() ?: break
+                            if (event.isBlank())
+                                continue;
+                            val eventJSON = event.removePrefix("data:")
+                            val data = Gson().fromJson(eventJSON, ChatCompletion::class.java);
+
+                            SwingUtilities.invokeLater {
+//                                val content = processor.process(data?.choices?.getOrNull(0)?.delta?.content?:"")
+                                sd.insertString(sd.length, data?.choices?.getOrNull(0)?.delta?.content?:"", null)
+                            }
+
+                        }
+
                     }
                 }
             }
 
-            private fun sendRequest(message: String): ChatCompletion {
-                val url = URL("https://genapi.ntq.ai/v1/chat/completions")
-                val conn = url.openConnection() as HttpURLConnection
-                conn.requestMethod = "POST"
-                conn.setRequestProperty("Content-Type", "application/json")
-                var authorizationKey = "Bearer " + MyPluginSettings.getInstance().apiKey;
-                conn.setRequestProperty("Authorization", authorizationKey)
-                conn.doOutput = true
-
-                val jsonRequest = PromptAction.JsonRequest(
-                        model = "ntq-coder",
-                        messages = listOf(PromptAction.Message("user", message)),
-                        max_tokens = 4096,
-                )
-
-                val requestBody = Gson().toJson(jsonRequest)
-                conn.outputStream.write(requestBody.toByteArray())
-
-                val responseCode = conn.responseCode
-                val response = if (responseCode == HttpURLConnection.HTTP_OK) {
-                    conn.inputStream.reader().readText()
-                } else {
-                    "Error: $responseCode"
-                }
-
-                conn.disconnect()
-                return Gson().fromJson(response, ChatCompletion::class.java)
+            private fun createConnectionSpecs(): List<ConnectionSpec>? {
+                val spec = ConnectionSpec.Builder(ConnectionSpec.COMPATIBLE_TLS)
+                    .tlsVersions(TlsVersion.TLS_1_2)
+                    .cipherSuites(
+                        CipherSuite.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
+                    )
+                    .build()
+                return listOf<ConnectionSpec>(spec)
             }
+//            private fun sendRequest(message: String): ChatCompletion {
+//                val url = URL("https://genapi.ntq.ai/v1/chat/completions")
+//                val conn = url.openConnection() as HttpURLConnection
+//                conn.requestMethod = "POST"
+//                conn.setRequestProperty("Content-Type", "application/json")
+//                var authorizationKey = "Bearer " + MyPluginSettings.getInstance().apiKey;
+//                conn.setRequestProperty("Authorization", authorizationKey)
+//                conn.doOutput = true
+//
+//                val jsonRequest = PromptAction.JsonRequest(
+//                        model = "ntq-coder",
+//                        messages = listOf(PromptAction.Message("user", message)),
+//                        max_tokens = 4096
+//                )
+//
+//                val requestBody = Gson().toJson(jsonRequest)
+//                conn.outputStream.write(requestBody.toByteArray())
+//
+//                val responseCode = conn.responseCode
+//                val response = if (responseCode == HttpURLConnection.HTTP_OK) {
+//                    conn.inputStream.reader().readText()
+//                } else {
+//                    "Error: $responseCode"
+//                }
+//
+//                conn.disconnect()
+//                return Gson().fromJson(response, ChatCompletion::class.java)
+//            }
 
         }
     }